# internal/api/nodes.go
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"log"
	"net/http"
	"net/url"
	"strconv"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"

	"github.com/pasarguard/panel/internal/domain"
	"github.com/pasarguard/panel/internal/nodeclient"
	"github.com/pasarguard/panel/internal/store"
)

func (a *API) clientForNode(n *domain.Node) *nodeclient.Client {
	return nodeclient.New(n.Address, n.MasterKey, n.CertPEM)
}

type registerNodeRequest struct {
	Name      string          `json:"name"`
	Address   string          `json:"address"`
	MasterKey string          `json:"master_key"`
	CoreKey   string          `json:"core_key"`         // PasarGuard core-management key (optional)
	CertPEM   string          `json:"cert_pem"`         // node's self-signed cert (PEM) to pin
	GRPCPort  int             `json:"grpc_port"`        // PasarGuard-compat gRPC port (default 62050)
	Config    json.RawMessage `json:"config,omitempty"` // optional fixed Xray config to push
	HostInfo  string          `json:"host_info"`
}

func (a *API) registerNode(w http.ResponseWriter, r *http.Request) {
	var req registerNodeRequest
	if err := decodeJSON(r, &req); err != nil {
		writeError(w, http.StatusBadRequest, "invalid json body")
		return
	}
	if strings.TrimSpace(req.Name) == "" || strings.TrimSpace(req.Address) == "" || req.MasterKey == "" {
		writeError(w, http.StatusBadRequest, "name, address and master_key are required")
		return
	}

	// Trust-on-first-use: if no cert was provided, try to fetch and pin the
	// node's current certificate automatically. If that fails, the node is still
	// registered (TLS verification is skipped for it).
	if strings.TrimSpace(req.CertPEM) == "" {
		if fetched, err := nodeclient.FetchCert(req.Address); err == nil {
			req.CertPEM = fetched
		}
	}

	node, err := a.store.CreateNode(req.Name, req.Address, req.MasterKey, req.CertPEM, string(req.Config), req.GRPCPort, req.CoreKey, req.HostInfo)
	if err != nil {
		writeError(w, http.StatusInternalServerError, "failed to create node")
		return
	}

	client := a.clientForNode(node)
	ctx, cancel := context.WithTimeout(r.Context(), 20*time.Second)
	defer cancel()

	if len(req.Config) > 0 {
		if err := client.ApplyConfig(ctx, string(req.Config)); err != nil {
			// Roll the row back. Leaving it behind strands a node record holding
			// a master key for a node that accepted nothing, and — since nothing
			// makes `address` unique — the operator's obvious next move (fix the
			// config, submit again) silently creates a duplicate node.
			if delErr := a.store.DeleteNode(node.ID); delErr != nil {
				log.Printf("register node %s: rollback after rejected config failed: %v", node.ID, delErr)
			}
			writeError(w, http.StatusBadGateway, "node rejected config: "+err.Error())
			return
		}
	}

	// Probe health so the registration reflects reality.
	if h, err := client.Health(ctx); err == nil {
		_ = a.store.UpdateNodeHealth(node.ID, "online", h.CoreVersion, time.Now().Unix())
		node.Status = "online"
		node.Version = h.CoreVersion
	} else {
		_ = a.store.UpdateNodeHealth(node.ID, "offline", "", time.Now().Unix())
		node.Status = "offline"
	}

	writeJSON(w, http.StatusCreated, node)
}

func (a *API) listNodes(w http.ResponseWriter, _ *http.Request) {
	nodes, err := a.store.ListNodes()
	if err != nil {
		writeError(w, http.StatusInternalServerError, "failed to list nodes")
		return
	}
	writeJSON(w, http.StatusOK, nodes)
}

func (a *API) nodeHealth(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	node, _ := a.getNodeOr404(w, id)
	if node == nil {
		return
	}

	ctx, cancel := context.WithTimeout(r.Context(), 15*time.Second)
	defer cancel()

	h, err := a.clientForNode(node).Health(ctx)
	if err != nil {
		_ = a.store.UpdateNodeHealth(id, "offline", "", time.Now().Unix())
		writeError(w, http.StatusBadGateway, "node unreachable: "+err.Error())
		return
	}
	_ = a.store.UpdateNodeHealth(id, "online", h.CoreVersion, time.Now().Unix())
	writeJSON(w, http.StatusOK, h)
}

func (a *API) getNodeOr404(w http.ResponseWriter, id string) (*domain.Node, error) {
	node, err := a.store.GetNode(id)
	if err != nil {
		writeError(w, http.StatusNotFound, "node not found")
		return nil, err
	}
	return node, nil
}

// nodeDetailResponse exposes the full connection details of a node for the
// operator: addresses, ports, credentials and the certificate (needed to
// register the node in a customer's or the operator's PasarGuard panel).
type nodeDetailResponse struct {
	ID          string `json:"id"`
	Name        string `json:"name"`
	Host        string `json:"host"`
	Address     string `json:"address"`      // HTTP control address (panel -> node)
	ServicePort int    `json:"service_port"` // HTTP control port
	GRPCAddress string `json:"grpc_address"` // customer's panel -> node
	GRPCPort    int    `json:"grpc_port"`
	Protocol    string `json:"protocol"`
	MasterKey   string `json:"master_key"`
	CoreKey     string `json:"core_key"`
	CertPEM     string `json:"cert_pem"`
	Status      string `json:"status"`
	Version     string `json:"version,omitempty"`
	LastSeenAt  int64  `json:"last_seen_at,omitempty"`
	CreatedAt   int64  `json:"created_at"`
	HostInfo    string `json:"host_info"`
	// CertPinned is false when no usable certificate is stored, meaning the
	// panel talks to this node WITHOUT verifying its identity while sending the
	// master key on every request. That state is otherwise indistinguishable
	// from a healthy node, so it is reported explicitly.
	CertPinned bool `json:"cert_pinned"`
}

// getNodeDetail returns a node's full connection info (host, service/gRPC ports,
// keys and certificate) so the operator can copy them and configure a panel.
func (a *API) getNodeDetail(w http.ResponseWriter, r *http.Request) {
	node, err := a.store.GetNode(chi.URLParam(r, "id"))
	if err != nil {
		writeError(w, http.StatusNotFound, "node not found")
		return
	}
	grpcPort := node.GRPCPort
	if grpcPort <= 0 {
		grpcPort = 62050
	}
	writeJSON(w, http.StatusOK, nodeDetailResponse{
		ID:          node.ID,
		Name:        node.Name,
		Host:        hostOf(node.Address),
		Address:     node.Address,
		ServicePort: servicePortOf(node.Address),
		GRPCAddress: grpcAddressFor(node.Address, grpcPort),
		GRPCPort:    grpcPort,
		Protocol:    "grpc",
		MasterKey:   node.MasterKey,
		CoreKey:     node.CoreKey,
		CertPEM:     node.CertPEM,
		Status:      node.Status,
		Version:     node.Version,
		LastSeenAt:  node.LastSeenAt,
		CreatedAt:   node.CreatedAt,
		HostInfo:    node.HostInfo,
		CertPinned:  nodeclient.IsPinned(node.CertPEM),
	})
}

// updateNodeRequest is a PATCH body: every field is optional and an OMITTED
// field leaves the stored value alone. Pointers make "absent" distinguishable
// from "set to empty" — with plain strings, a client sending only the fields it
// wanted to change silently erased the others.
type updateNodeRequest struct {
	Name      *string `json:"name,omitempty"`
	Address   *string `json:"address,omitempty"`
	GRPCPort  *int    `json:"grpc_port,omitempty"`
	MasterKey *string `json:"master_key,omitempty"`
	CoreKey   *string `json:"core_key,omitempty"`
	CertPEM   *string `json:"cert_pem,omitempty"`
	HostInfo  *string `json:"host_info,omitempty"`
}

// updateNode applies a partial edit to a node's name/address/ports/keys/cert.
func (a *API) updateNode(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	var req updateNodeRequest
	if err := decodeJSON(r, &req); err != nil {
		writeError(w, http.StatusBadRequest, "invalid json body")
		return
	}
	// Name and address may be omitted, but not blanked: a node with neither is
	// unusable and unreachable.
	if req.Name != nil && strings.TrimSpace(*req.Name) == "" {
		writeError(w, http.StatusBadRequest, "name cannot be empty")
		return
	}
	if req.Address != nil && strings.TrimSpace(*req.Address) == "" {
		writeError(w, http.StatusBadRequest, "address cannot be empty")
		return
	}
	if err := a.store.UpdateNode(id, store.NodeUpdate{
		Name:      req.Name,
		Address:   req.Address,
		GRPCPort:  req.GRPCPort,
		MasterKey: req.MasterKey,
		CoreKey:   req.CoreKey,
		CertPEM:   req.CertPEM,
		HostInfo:  req.HostInfo,
	}); err != nil {
		writeNotFoundOr500(w, err)
		return
	}
	node, err := a.store.GetNode(id)
	if err != nil {
		writeNotFoundOr500(w, err)
		return
	}
	writeJSON(w, http.StatusOK, node)
}

// servicePortOf extracts the HTTP control port from a node address (default 8090).
func servicePortOf(address string) int {
	if u, err := url.Parse(address); err == nil && u.Port() != "" {
		if p, err := strconv.Atoi(u.Port()); err == nil {
			return p
		}
	}
	return 8090
}

// coreLifecycle handles start/stop/restart of a node's shared core.
func (a *API) coreLifecycle(w http.ResponseWriter, r *http.Request) {
	node, err := a.store.GetNode(chi.URLParam(r, "id"))
	if err != nil {
		writeError(w, http.StatusNotFound, "node not found")
		return
	}
	action := chi.URLParam(r, "action")
	ctx, cancel := context.WithTimeout(r.Context(), 30*time.Second)
	defer cancel()
	cl := a.clientForNode(node)
	switch action {
	case "start":
		err = cl.CoreStart(ctx)
	case "stop":
		err = cl.CoreStop(ctx)
	case "restart":
		err = cl.CoreRestart(ctx)
	default:
		writeError(w, http.StatusBadRequest, "unknown action")
		return
	}
	if err != nil {
		writeError(w, http.StatusBadGateway, "node error: "+err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]string{"status": action})
}

type versionRequest struct {
	Version string `json:"version"`
}

// setXrayVersion asks the node to switch its Xray-core version and restart.
func (a *API) setXrayVersion(w http.ResponseWriter, r *http.Request) {
	node, err := a.store.GetNode(chi.URLParam(r, "id"))
	if err != nil {
		writeError(w, http.StatusNotFound, "node not found")
		return
	}
	var req versionRequest
	// An absent body means "latest". A malformed one is rejected rather than
	// silently treated as empty, which would push a different version onto the
	// node than the operator asked for.
	if err := decodeJSON(r, &req); err != nil && !errors.Is(err, io.EOF) {
		writeError(w, http.StatusBadRequest, "invalid json body")
		return
	}
	ctx, cancel := context.WithTimeout(r.Context(), 190*time.Second)
	defer cancel()
	if err := a.clientForNode(node).SetXrayVersion(ctx, req.Version); err != nil {
		writeError(w, http.StatusBadGateway, "node error: "+err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]string{"status": "xray updated"})
}

// updateNodeBinary asks the node to self-update its binary and restart.
func (a *API) updateNodeBinary(w http.ResponseWriter, r *http.Request) {
	node, err := a.store.GetNode(chi.URLParam(r, "id"))
	if err != nil {
		writeError(w, http.StatusNotFound, "node not found")
		return
	}
	var req versionRequest
	if err := decodeJSON(r, &req); err != nil {
		writeError(w, http.StatusBadRequest, "invalid json body")
		return
	}
	ctx, cancel := context.WithTimeout(r.Context(), 190*time.Second)
	defer cancel()
	if err := a.clientForNode(node).UpdateNode(ctx, req.Version); err != nil {
		writeError(w, http.StatusBadGateway, "node error: "+err.Error())
		return
	}
	writeJSON(w, http.StatusOK, map[string]string{"status": "updating"})
}

// deleteNode removes a node. Nodes that still have subscriptions cannot be
// deleted (delete the subscriptions first) — otherwise the FK reference would
// surface as a raw 500 and, worse, live tenants would be left unmanaged.
func (a *API) deleteNode(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	// Fail closed: a DB error here must not skip the guard and let the delete
	// proceed against a node that may well still have live subscriptions.
	n, err := a.store.CountSubscriptionsByNode(id)
	if err != nil {
		writeError(w, http.StatusInternalServerError, "failed to check node subscriptions: "+err.Error())
		return
	}
	if n > 0 {
		writeError(w, http.StatusConflict,
			fmt.Sprintf("node has %d subscription(s); delete them first", n))
		return
	}
	if err := a.store.DeleteNode(id); err != nil {
		writeNotFoundOr500(w, err)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

// getNodeConfig returns the node's currently running Xray config. It queries the
// node live (so the GUI/operator sees the real inbound to share with customers)
// and falls back to the last config the panel pushed if the node is unreachable.
func (a *API) getNodeConfig(w http.ResponseWriter, r *http.Request) {
	node, err := a.store.GetNode(chi.URLParam(r, "id"))
	if err != nil {
		writeError(w, http.StatusNotFound, "node not found")
		return
	}
	ctx, cancel := context.WithTimeout(r.Context(), 15*time.Second)
	defer cancel()

	w.Header().Set("Content-Type", "application/json")
	if live, err := a.clientForNode(node).GetConfig(ctx); err == nil && len(bytes.TrimSpace(live)) > 0 {
		_, _ = w.Write(live)
		return
	}
	if node.ConfigJSON == "" {
		_, _ = w.Write([]byte("{}"))
		return
	}
	_, _ = w.Write([]byte(node.ConfigJSON))
}

// getNodeInbounds returns the customer-shareable inbound definitions of a node
// ({"inbounds":[...]}). Prefers the live node; falls back to the panel's stored
// config so it still works with an older/unreachable node.
func (a *API) getNodeInbounds(w http.ResponseWriter, r *http.Request) {
	node, err := a.store.GetNode(chi.URLParam(r, "id"))
	if err != nil {
		writeError(w, http.StatusNotFound, "node not found")
		return
	}
	ctx, cancel := context.WithTimeout(r.Context(), 15*time.Second)
	defer cancel()
	w.Header().Set("Content-Type", "application/json")
	_, _ = w.Write(a.resolveInbounds(ctx, node))
}

// resolveInbounds returns a node's customer-shareable inbounds. It prefers the
// live node (which redacts secrets and adds the Reality public key); if the node
// is unreachable or older (no /admin/inbounds) or returns nothing, it falls back
// to the panel's stored fixed config with the Reality private key stripped.
func (a *API) resolveInbounds(ctx context.Context, node *domain.Node) json.RawMessage {
	if ib, err := a.clientForNode(node).GetInbounds(ctx); err == nil && hasInbounds(ib) {
		return ib
	}
	if fb := inboundsFromConfig(node.ConfigJSON); fb != nil {
		return fb
	}
	return json.RawMessage(`{"inbounds":[]}`)
}

func hasInbounds(raw json.RawMessage) bool {
	var d struct {
		Inbounds []json.RawMessage `json:"inbounds"`
	}
	if json.Unmarshal(raw, &d) != nil {
		return false
	}
	return len(d.Inbounds) > 0
}

// inboundsFromConfig extracts {"inbounds":[...]} from a full Xray config and
// strips server-only secrets (Reality privateKey) before sharing.
func inboundsFromConfig(cfg string) json.RawMessage {
	if cfg == "" {
		return nil
	}
	var doc struct {
		Inbounds []json.RawMessage `json:"inbounds"`
	}
	if json.Unmarshal([]byte(cfg), &doc) != nil || len(doc.Inbounds) == 0 {
		return nil
	}
	cleaned := make([]json.RawMessage, 0, len(doc.Inbounds))
	for _, ib := range doc.Inbounds {
		cleaned = append(cleaned, redactInboundJSON(ib))
	}
	out, err := json.Marshal(struct {
		Inbounds []json.RawMessage `json:"inbounds"`
	}{Inbounds: cleaned})
	if err != nil {
		return nil
	}
	return out
}

func redactInboundJSON(raw json.RawMessage) json.RawMessage {
	var m map[string]any
	if json.Unmarshal(raw, &m) != nil {
		return raw
	}
	if ss, ok := m["streamSettings"].(map[string]any); ok {
		if rs, ok := ss["realitySettings"].(map[string]any); ok {
			delete(rs, "privateKey")
		}
	}
	out, err := json.Marshal(m)
	if err != nil {
		return raw
	}
	return out
}

// updateNodeConfig pushes a new fixed Xray config to the node and stores it.
func (a *API) updateNodeConfig(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	node, err := a.store.GetNode(id)
	if err != nil {
		writeError(w, http.StatusNotFound, "node not found")
		return
	}

	body, err := io.ReadAll(io.LimitReader(r.Body, 8<<20))
	if err != nil || len(bytes.TrimSpace(body)) == 0 {
		writeError(w, http.StatusBadRequest, "config body is empty")
		return
	}
	if !json.Valid(body) {
		writeError(w, http.StatusBadRequest, "config must be valid JSON")
		return
	}

	ctx, cancel := context.WithTimeout(r.Context(), 30*time.Second)
	defer cancel()
	if err := a.clientForNode(node).ApplyConfig(ctx, string(body)); err != nil {
		writeError(w, http.StatusBadGateway, "node rejected config: "+err.Error())
		return
	}
	if err := a.store.UpdateNodeConfig(id, string(body)); err != nil {
		writeError(w, http.StatusInternalServerError, "failed to store config")
		return
	}

	if h, err := a.clientForNode(node).Health(ctx); err == nil {
		_ = a.store.UpdateNodeHealth(id, "online", h.CoreVersion, time.Now().Unix())
	}
	writeJSON(w, http.StatusOK, map[string]string{"status": "applied"})
}