# internal/web/assets/openapi.yaml
openapi: 3.0.3
info:
  title: PasarGuard Panel API
  version: "1.0.0"
  description: >
    API پنل اصلی برای مدیریت چند نود، مشتری، پلن، اشتراک و سهمیه.
    این API سطح یکپارچه‌سازی برای ربات فروش است. پنل منطق پول/کیف‌پول ندارد؛
    فقط بایت/زمان/وضعیت را مدیریت و مصرف/overage را expose می‌کند.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/nodes/{id}:
    get:
      tags: [nodes]
      summary: Node connection details (host, ports, keys, certificate) for the operator
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Node detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeDetail" }
    patch:
      tags: [nodes]
      summary: Update a node (partial; omitted fields keep their stored value)
      description: >
        هر فیلد اختیاری است. فیلدی که ارسال نشود بدون تغییر می‌ماند؛ ارسال رشته‌ی
        خالی یعنی «پاک کن». name و address را می‌توان حذف کرد ولی نمی‌توان خالی فرستاد.
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                address: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                grpc_port: { type: integer, description: "ارسال‌نشده = بدون تغییر" }
                master_key: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                core_key: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                cert_pem: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                host_info: { type: string, description: "ارسال‌نشده = بدون تغییر" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Node" } } } }
    delete:
      tags: [nodes]
      summary: Delete a node
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    Node:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        address: { type: string }
        grpc_port: { type: integer, example: 62050 }
        status: { type: string, enum: [online, offline, unknown] }
        version: { type: string }
        capacity_score: { type: integer }
        last_seen_at: { type: integer, format: int64 }
        created_at: { type: integer, format: int64 }
    NodeDetail:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        host: { type: string, example: "1.2.3.4" }
        address: { type: string, example: "https://1.2.3.4:8090" }
        service_port: { type: integer, example: 8090 }
        grpc_address: { type: string, example: "1.2.3.4:62050" }
        grpc_port: { type: integer, example: 62050 }
        protocol: { type: string, example: "grpc" }
        master_key: { type: string }
        core_key: { type: string }
        cert_pem: { type: string }
        cert_pinned:
          type: boolean
          description: >
            false یعنی گواهی معتبری ذخیره نشده و ارتباط پنل با نود بدون تأیید هویت
            انجام می‌شود (در حالی که master key در هر درخواست ارسال می‌گردد).
        host_info:
          type: string
          description: "یادداشت آزاد اپراتور؛ در پاسخ connection به خریدار داده می‌شود"
        status: { type: string }
        version: { type: string }
        last_seen_at: { type: integer, format: int64 }
        created_at: { type: integer, format: int64 }
...